+++ js/src/main/scala/zio/pdf/PdfEngine.scala
@@ -491,3 +491,10 @@ object PdfEngine:
     opts: Options = Options.default
   ): ZIO[PdfEngine, Throwable, Chunk[Byte]] =
     PdfSplit.rotateBytes(bytes, degrees, fromPage, toPage, opts)
+
+  def watermark(
+    bytes: Chunk[Byte],
+    stamp: PdfWatermark.Text,
+    opts: Options = Options.default
+  ): ZIO[PdfEngine, Throwable, Chunk[Byte]] =
+    PdfWatermark.fromBytes(bytes, stamp, opts)